requirements.txt
pandas==2.0.1
SQLAlchemy==2.0.12
psycopg2-binary==2.9.6